Replace style sharing cousin list search with LRU
The current implementation of style sharing uses a very confusing set of methods to
search a node's cousins in an attempt to find other nodes to share style with. This
patch replaces this confusing method with a simple LRU that allows us to find
nodes to share with even if they're not our cousins.

The previous implementation of this was reverted due to a performance regression
due to decreased sharing occurring in certain situations during parse time when
we'd clear the LRU after every node was appended. This updated patch uses RefPtrs
to allow these nodes to be held onto, and avoids clearing the LRU during attach,
and only when styles change.

Because this change introduces RefPtrs that can hold onto nodes while the style
resolver is active, Internal gc functions are updated to clear the style sharing list.

Previous codereview: https://codereview.chromium.org/23075004/

Review URL: https://chromiumcodereview.appspot.com/23049008

git-svn-id: svn://svn.chromium.org/blink/trunk@157311 bbb929c8-8fbe-4397-9dbb-9b2b20218538