[CSSRegions] Pseudo-elements as regions should not be exposed to JS
At this moment, it is possible to expose a pseudo-element from a region in JS
by means of NamedFlow::getRegions() and NamedFlow::getRegionsByContent() API.
We should do that only after we properly implement the Region interface.
In addition, the patch skips the pseudo-elements as regions when computing
the first empty region index since we do not want to return an index that cannot
be used with the list of regions returned by NamedFlow::getRegions() api.

BUG=285317

Review URL: https://chromiumcodereview.appspot.com/23622019

git-svn-id: svn://svn.chromium.org/blink/trunk@157317 bbb929c8-8fbe-4397-9dbb-9b2b20218538