Make it possible to focus on 0-size form controls and anchors.
We assumed form controls and anchor elements with 0-size were not
focusable. It's inconsistent with other elements with tabIndex or
contentEditable, and incompatible with IE11 and Firefox 23.

We remove rendererIsFocusable override of HTMLFormControlElement, and
remove renderer box check in HTMLAnchorElement::isKeyboardFocusable.

As for crbug.com/285595, when we removed focus from an input/textarea element by
0-size, the element kept the selection. So FrameSelection::setFocusedNodeIfNeeded
set focus on the element again when an editing command was invoked by a keyboard
input. This issue is revolved by this CL because we don't remove focus by
0-size.

BUG=168121,285595

Review URL: https://chromiumcodereview.appspot.com/23496036

git-svn-id: svn://svn.chromium.org/blink/trunk@157405 bbb929c8-8fbe-4397-9dbb-9b2b20218538