Fix crash in OverlayerGraphicsLayerClientImpl::debugName
This method was bound to crash. It was doing a pointer comparison with a value
that must be NULL. Updated the code so that the client retains a pointer to
the graphics layer it creates for later comparison with the graphics layer it
receives in ::debugName.

R=danakj
BUG=None

Review URL: https://chromiumcodereview.appspot.com/23902013

git-svn-id: svn://svn.chromium.org/blink/trunk@157606 bbb929c8-8fbe-4397-9dbb-9b2b20218538