Permanently enable AXComputedObjectAttributeCache.
It was never needed by other WebKit ports, but it's always a
performance win for Chromium. It's actually better performance to
just leave it enabled than to wait for Chromium to enable it.

After this lands, we can stop calling the functions to enable and
disable it from Chromium and then remove the API.

BUG=289689

Review URL: https://chromiumcodereview.appspot.com/23648013

git-svn-id: svn://svn.chromium.org/blink/trunk@157773 bbb929c8-8fbe-4397-9dbb-9b2b20218538