Fix crash bug where overlap map was accesing below the bottom of its stack
A crash occurs with an unlucky combination of (1) compositing not
enabled yet, (2) composited negative z-index child, and (3) a second,
non-composited negative z-index child. In this scenario, the second
child was trying to add its bounds to the overlap map but no one
had actually created an overlap context on the stack.

The rule of thumb is that whenever we defined a layer as compositing
in computeCompositingRequirements, we must make sure we created
a new overlap context for its subtree. This was not happening
for the composited negative z-index child.

BUG=285979

Review URL: https://chromiumcodereview.appspot.com/23498037

git-svn-id: svn://svn.chromium.org/blink/trunk@157803 bbb929c8-8fbe-4397-9dbb-9b2b20218538