ASSERTION FAILED: !object || object->isBox(), UNKNOWN in WebCore::RenderNamedFlowThread::getRanges
The inline SVG element behaves like a replaced element in HTML context. When computing the ranges
that are displayed inside a region for a named flow that contains an inline SVG element, we have
to treat the SVG as a replaced and run the checks only for the SVG root element, skipping the
other SVG elements inside.

I have added a new function, nextNodeInsideContentNode, that is used to retrieve the next node
under the contentNode tree that should be used for getRanges inspection. In the future, further
restrictions can be added to this function if needed.

Also made some changes to make sure we do not have blink coding style errors.

BUG=291829

Review URL: https://chromiumcodereview.appspot.com/23549035

git-svn-id: svn://svn.chromium.org/blink/trunk@157959 bbb929c8-8fbe-4397-9dbb-9b2b20218538