Prefer jinja2 from third_party over system library in bindings too
* Move third_party to front of sys.path - insert(1, ...) - instead of end
  (.extend(...))
* Also move insertion to position 1, not 0.

sys.path[0] is supposed to be directory containing the script (or current
directory), so insert *after* this to not break this assumption.

Per docs:
http://docs.python.org/2/library/sys.html#sys.path

As initialized upon program startup, the first item of this list, path[0], is the directory containing the script that was used to invoke the Python interpreter.

See discussion:
http://stackoverflow.com/questions/10095037/why-use-sys-path-appendpath-instead-of-sys-path-insert1-path


Followup to:
Prefer jinja2 from third_party over eventual system jinja2 libraries.
https://chromiumcodereview.appspot.com/23498033/

Review URL: https://chromiumcodereview.appspot.com/24085005

git-svn-id: svn://svn.chromium.org/blink/trunk@157960 bbb929c8-8fbe-4397-9dbb-9b2b20218538