Always send the current URL/title to the activity log when logging
When calling into WebDOMActivityLogger::log, the current location and
title for the page should always be passed in--otherwise, when a
WebDOMActivityLogger is re-used across multiple pages stale data can be
logged.

This is the first in a sequence of changes, and defines and calls a new
logging API in the blink tree.  Actual logging will still go through the
old path until the corresponding new method is added in the Chrome tree.

BUG=294500

Review URL: https://chromiumcodereview.appspot.com/23817008

git-svn-id: svn://svn.chromium.org/blink/trunk@158128 bbb929c8-8fbe-4397-9dbb-9b2b20218538