Move tests to new import location
The import script was previously importing the tests into an incorrect
path. This change uses the latest import script to import all of the
shapes tests to the right location.

There are also some small changes in the tests since they had some more
metadata added upstream, but there are no functional changes in the
tests with this change.

Review URL: https://chromiumcodereview.appspot.com/24191009

git-svn-id: svn://svn.chromium.org/blink/trunk@158193 bbb929c8-8fbe-4397-9dbb-9b2b20218538