Stop processing bytes after a DocumentParser is stopped or detached
It's possible to detach a Document from its DocumentParser before all
bytes for the resource are received if the page calls document.open() or
document.write() without a defined insertion point. If that happens, the
original document is blown away and further bytes from the original
resource should just be thrown away.

This is a regression from somewhere around r152864

BUG=297359

Review URL: https://chromiumcodereview.appspot.com/24433002

git-svn-id: svn://svn.chromium.org/blink/trunk@158300 bbb929c8-8fbe-4397-9dbb-9b2b20218538